refactor: move core.bufline, core.dirty_range -> parser.bufline, parser.dirty_range
