+++ lua/tirenvi/app/pipeline.lua
@@ -6,8 +6,8 @@ local Document = require("tirenvi.core.document")
 local Attrs = require("tirenvi.core.attrs")
 local Attr = require("tirenvi.core.attr")
 local Cell = require("tirenvi.core.cell")
-local Bufline = require("tirenvi.core.bufline")
-local dirty_range = require("tirenvi.core.dirty_range")
+local Bufline = require("tirenvi.parser.bufline")
+local dirty_range = require("tirenvi.parser.dirty_range")
 local Request = require("tirenvi.app.request")
 local width_layout = require("tirenvi.width.layout")
 local flat_parser = require("tirenvi.parser.flat_parser")
