+++ lua/tirenvi/editor/debug.lua
@@ -2,7 +2,7 @@
 local Context = require("tirenvi.app.context")
 local Attrs = require("tirenvi.core.attrs")
 local Attr = require("tirenvi.core.attr")
-local bufline = require("tirenvi.core.bufline")
+local bufline = require("tirenvi.parser.bufline")
 local CursorLogical = require("tirenvi.cursor.cursor_logical")
 local CursorNvim = require("tirenvi.cursor.nvim")
 local CursorConvert = require("tirenvi.cursor.convert")
