+++ lua/tirenvi/init.lua
@@ -6,7 +6,7 @@ local buffer = require("tirenvi.io.buffer")
 local reader = require("tirenvi.io.reader")
 local buf_state = require("tirenvi.io.buf_state")
 local attr_store = require("tirenvi.io.attr_store")
-local Bufline = require("tirenvi.core.bufline")
+local Bufline = require("tirenvi.parser.bufline")
 local Parser = require("tirenvi.parser.parser")
 local Range = require("tirenvi.util.range")
 local util = require("tirenvi.util.util")
